Add H10 and H12 probing pipelines with run_slug output layout.
H10 joins external annotations by example_id; H12 decodes abstain_logit on
with_abstain choice rows. Shared core/load_run updates and expanded probes README.

+++ probes/load_run.py
@@ -45,7 +45,8 @@ def load_hidden_states(run_dir: Path) -> tuple[np.ndarray, list[str]]:
             "Download from HF: python -m probes.download_hf"
         )
     data = np.load(path)
-    hs = np.asarray(data["hs"], dtype=np.float32)
+    # Keep source dtype (float16 in canonical runs) to avoid a full-run float32 copy.
+    hs = np.asarray(data["hs"])
     item_ids = [str(x) for x in data["item_ids"]]
     return hs, item_ids
